refactor(data): rename generate-qr to update-landmarks-data
